Added mixin for used scripts
