+++ app/src/components/WalletHistory/TxHistoryMixin.js
@@ -1,0 +1,64 @@
+const {
+  ROPSTEN,
+  RINKEBY,
+  KOVAN,
+  MAINNET,
+  LOCALHOST,
+  GOERLI,
+  RPC,
+  ROPSTEN_DISPLAY_NAME,
+  RINKEBY_DISPLAY_NAME,
+  KOVAN_DISPLAY_NAME,
+  MAINNET_DISPLAY_NAME,
+  LOCALHOST_DISPLAY_NAME,
+  GOERLI_DISPLAY_NAME,
+  RPC_DISPLAY_NAME
+} = require('../../utils/enums')
+
+const mapper = {
+  [ROPSTEN]: ROPSTEN_DISPLAY_NAME,
+  [RINKEBY]: RINKEBY_DISPLAY_NAME,
+  [KOVAN]: KOVAN_DISPLAY_NAME,
+  [MAINNET]: MAINNET_DISPLAY_NAME,
+  [LOCALHOST]: LOCALHOST_DISPLAY_NAME,
+  [GOERLI]: GOERLI_DISPLAY_NAME,
+  [RPC]: RPC_DISPLAY_NAME
+}
+
+export default {
+  props: ['transactions', 'selectedAction', 'selectedPeriod'],
+  data() {
+    return {
+      mapper: mapper
+    }
+  },
+  computed: {
+    filteredTransactions() {
+      const selectedAction = this.selectedAction === 'All Transactions' ? '' : this.selectedAction
+      var regExAction = new RegExp(selectedAction, 'i')
+
+      return this.transactions.filter(item => {
+        // GET Date Scope
+        let isScoped = false
+        if (this.selectedPeriod === 'Period') {
+          isScoped = true
+        } else {
+          let minDate = new Date()
+          let itemDate = new Date(item.date)
+          if (this.selectedPeriod === 'Last Week') {
+            minDate.setDate(minDate.getDate() - 7)
+          } else {
+            minDate.setMonth(minDate.getMonth() - 1)
+          }
+
+          isScoped = minDate.getTime() <= itemDate.getTime()
+        }
+        if (item.action) {
+          return item.action.match(regExAction) && isScoped
+        } else {
+          return isScoped
+        }
+      })
+    }
+  }
+}
