+++ app/src/components/WalletHistory/TxHistoryTableMobile/TxHistoryTableMobile.vue
@@ -81,112 +81,21 @@
 </template>
 
 <script>
-const {
-  ROPSTEN,
-  RINKEBY,
-  KOVAN,
-  MAINNET,
-  LOCALHOST,
-  GOERLI,
-  RPC,
-  ROPSTEN_DISPLAY_NAME,
-  RINKEBY_DISPLAY_NAME,
-  KOVAN_DISPLAY_NAME,
-  MAINNET_DISPLAY_NAME,
-  LOCALHOST_DISPLAY_NAME,
-  GOERLI_DISPLAY_NAME,
-  RPC_DISPLAY_NAME
-} = require('../../../utils/enums')
-
-const mapper = {
-  [ROPSTEN]: ROPSTEN_DISPLAY_NAME,
-  [RINKEBY]: RINKEBY_DISPLAY_NAME,
-  [KOVAN]: KOVAN_DISPLAY_NAME,
-  [MAINNET]: MAINNET_DISPLAY_NAME,
-  [LOCALHOST]: LOCALHOST_DISPLAY_NAME,
-  [GOERLI]: GOERLI_DISPLAY_NAME,
-  [RPC]: RPC_DISPLAY_NAME
-}
+import TxHistoryMixin from '../TxHistoryMixin'
 
 export default {
-  props: ['transactions', 'selectedAction', 'selectedPeriod'],
+  mixins: [TxHistoryMixin],
   data() {
     return {
-      expanded: [],
       expand: false,
       pagination: {},
       defaultSort: 'date',
-      search: '',
-      headers: [
-        {
-          text: 'Transaction',
-          value: 'action',
-          align: 'left',
-          width: '150px'
-        },
-        {
-          text: 'From',
-          value: 'from',
-          align: 'left',
-          width: '350px'
-        },
-        {
-          text: 'To',
-          value: 'to',
-          align: 'left',
-          width: '350px'
-        },
-        {
-          text: 'Amount',
-          value: 'amount',
-          align: 'right',
-          width: '200px'
-        },
-        {
-          text: 'Date',
-          value: 'date',
-          align: 'right',
-          width: '80px'
-        },
-        {
-          text: 'Status',
-          value: 'status',
-          align: 'center'
-        }
-      ],
-      mapper: mapper
+      search: ''
     }
   },
   computed: {
     showFooter() {
       return this.transactions && this.transactions.length > 5
-    },
-    filteredTransactions() {
-      const selectedAction = this.selectedAction === 'All Transactions' ? '' : this.selectedAction
-      var regExAction = new RegExp(selectedAction, 'i')
-
-      return this.transactions.filter(item => {
-        // GET Date Scope
-        let isScoped = false
-        if (this.selectedPeriod === 'Period') {
-          isScoped = true
-        } else {
-          let minDate = new Date()
-          let itemDate = new Date(item.date)
-          if (this.selectedPeriod === 'Last Week') {
-            minDate.setDate(minDate.getDate() - 7)
-          } else {
-            minDate.setMonth(minDate.getMonth() - 1)
-          }
-
-          isScoped = minDate.getTime() <= itemDate.getTime()
-        }
-        if (item.action) {
-          return item.action.match(regExAction) && isScoped
-        } else {
-          return isScoped
-        }
-      })
     }
   },
   methods: {
@@ -196,13 +105,6 @@ export default {
       } else {
         this.pagination.sortBy = column
         this.pagination.descending = false
-      }
-    },
-    rowClicked(item) {
-      if (this.expanded.indexOf(item) >= 0) {
-        this.expanded = []
-      } else {
-        this.expanded = [item]
       }
     },
     formatDate(date) {
